Check that test user has access to cluster before proceeding in tests
